+++ commands/game.js
@@ -4,11 +4,16 @@ exports.run = async (bot, message, args) => { // eslint-disable-line no-unused-v
     if (bot.activeGames[message.guild.id] && bot.activeGames[message.guild.id][message.channel.id]) {
         if (!(bot.activeGames[message.guild.id][message.channel.id] instanceof Game))
             return message.channel.send("No active game detected in this channel!");
-        if (!(bot.activeGames[message.guild.id][message.channel.id].some(element => element.id === message.author.id)))
+        if (!(bot.activeGames[message.guild.id][message.channel.id].players.some(element => element.id === message.author.id)))
             return message.channel.send("Only active members of the active game may use this command.");
 
         if (args[0] === "resend") {
             bot.activeGames[message.guild.id][message.channel.id].resend();
+            return;
+        }
+        if (args[0] === "abort") {
+            bot.activeGames[message.guild.id][message.channel.id].abortGame(message.author);
+            return;
         }
     }
     else return message.channel.send("No active game detected in this channel!");
@@ -25,5 +30,5 @@ exports.conf = {
 exports.help = {
     name: "game",
     description: "Manage the game instance running in the server, such as resending game messages and canceling.",
-    usage: "game [resend]"
+    usage: "game [resend/abort]"
 };