+++ commands/setup.js
@@ -1,13 +1,34 @@
 const GameSetup = require('../classes/GameSetup.js');
-
 exports.run = async (bot, message, args) => { // eslint-disable-line no-unused-vars
     if (bot.activeGames[message.guild.id] && bot.activeGames[message.guild.id][message.channel.id]) {
-        if (bot.activeGames[message.guild.id][message.channel.id].gm.id !== message.author.id) {
-            return message.channel.send(`${message.author} Only the host of the game may set it up!`);
-        } else {
-            bot.activeGames[message.guild.id][message.channel.id].interpretCommand(message, args);
+        // bot.logger.log('debug', bot.activeGames[message.guild.id][message.channel.id] instanceof Game);
+        if (bot.activeGames[message.guild.id][message.channel.id] instanceof GameSetup) {
+            if (args[0].match(/^(leave|resend)$/i) && !bot.activeGames[message.guild.id][message.channel.id].players.some(e => e.id == message.author.id))
+                return message.channel.send(`${message.author} You must be in the game to make changes to it!`);
+            if (bot.activeGames[message.guild.id][message.channel.id].gm.id !== message.author.id)
+                return message.channel.send(`${message.author} Only the host of the game may set it up!`);
+            else
+                bot.activeGames[message.guild.id][message.channel.id].interpretCommand(message, args);
+        } else { // Implicit assumption that if it's not GameSetup then it must be an instance of Game
+            if (args[0] === 'redo') {
+                if (bot.activeGames[message.guild.id][message.channel.id].timeout)
+                    bot.activeGames[message.guild.id][message.channel.id].renewGame(message);
+                else {
+                    let temp = await message.channel.send("The game running in this channel hasn't completed yet!");
+                    temp.delete({ timeout: 3000 });
+                    return;
+                }
+            }
+            else {
+                let temp = await message.channel.send("A game is already running in this channel!");
+                temp.delete({ timeout: 3000 });
+                return;
+            }
         }
     } else {
+        if (bot.activeGames[message.guild.id] && Object.keys(bot.activeGames[message.guild.id]).length >= bot.config.gamesPerSever)
+            return message.channel.send(`The game limit per guild has been hit. Please wait for an existing game to end before trying again. (limit: ${bot.config.gamesPerSever})`);
+
         // Check if game exists
         let gameIndex = bot.games.findIndex(element => {
             let g = args.join(' ').toLowerCase();
@@ -21,6 +42,7 @@ exports.run = async (bot, message, args) => { // eslint-disable-line no-unused-v
 
         if (!bot.activeGames[message.guild.id]) bot.activeGames[message.guild.id] = {};
         if (!bot.activeGames[message.guild.id][message.channel.id]) bot.activeGames[message.guild.id][message.channel.id] = new GameSetup(bot, message, bot.games[gameIndex]);
+        await bot.activeGames[message.guild.id][message.channel.id].init();
     }
 };
 
@@ -35,6 +57,6 @@ exports.conf = {
 exports.help = {
     name: "setup",
     description: "Starts a game setup menu, allowing you to invite other players or set game options. Once setup has started, you can also use this command to invite players and set up game options.",
-    usage: "setup [game] OR setupgame [invite/option/turnorder/resend/start/cancel] [...]",
-    examples: ["setup Tic-tac-toe", "setup invite @User#1234 @User2#5678", "setup turnorder 1 @User#1234"]
+    usage: "setup [game] OR setupgame [invite/option/turns/resend/leave/kick/start/cancel] [...]",
+    examples: ["setup Tic-tac-toe", "setup invite @User#1234 @User2#5678", "setup turns 1 @User#1234"]
 };