+++ index.js
@@ -23,7 +23,7 @@ bot.games = [];
 
 /*
 There's probably a better way to do this, but enmaps can't store classes
-so I'm just using an object to store active games with some additional GC code (see ./common/bot_functions.js)
+so I'm just using an object to store active games with some additional GC code
 I'm not sure if the GC code is necessary but better safe than memory leak down the line
 */
 bot.activeGames = {};